requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
tensorboardX>=2.6.0
Pillow>=10.0.0
tqdm>=4.66.0
